pyproject.toml
[project]
name = "mosektools"
version = "0.0.0"
description = "..."
authors = [{name = "Thomas Schmelzer", email = "thomas.schmelzer@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mosek>=10.2.12",
    "numpy>=2.2.1",
]

[project.urls]
repository = "https://github.com/tschm/MosekRegression"

[project.optional-dependencies]
dev = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.3",
    "pre-commit>=4.0.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "yfinance>=0.2.52",
]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    "*__init__.py"
]

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mosek_tools"]

[tool.bandit]
exclude_dirs = [".venv", "tests"]

[tool.deptry]
# see https://deptry.com/usage/#pep-621-dev-dependency-groups
pep621_dev_dependency_groups = ["dev"]
